Convert all our protocol lengths to uint32_t.
This makes the protocol compatible between 32/64 bit
and locks down all our sizes. Internally we still use
size_t everywhere so caution must be used with esp.
memcpy - avoid using this and use casts instead.

+++ src/include/bt-daemon.h
@@ -32,6 +32,7 @@
 
 #include <stdbool.h>
 #include <unistd.h>
+#include <stdint.h>
 #include <sys/types.h>
 
 #if (__GNUC__ >= 4)
@@ -70,7 +71,7 @@ typedef enum BuxtonDataType {
  */
 typedef struct BuxtonString {
 	char *value; /**<The content of the string */
-	size_t length; /**<The length of the string */
+	uint32_t length; /**<The length of the string */
 } BuxtonString;
 
 /**
