+++ src/shared/util.h
@@ -52,7 +52,7 @@
 size_t page_size(void);
 #define PAGE_ALIGN(l) ALIGN_TO((l), page_size())
 
-#define buxton_string_pack(s) ((BuxtonString){(s), strlen(s) + 1})
+#define buxton_string_pack(s) ((BuxtonString){(s), (uint32_t)strlen(s) + 1})
 
 #define streq(a,b) (strcmp((a),(b)) == 0)
 #define strneq(a, b, n) (strncmp((a), (b), (n)) == 0)
